# docs/openapi/flagman-delivery.openapi.yaml
openapi: 3.0.3
info:
  title: Flagman Delivery API
  version: 1.0.0
  description: |
    OpenAPI specification generated from the FastAPI backend in `hac.new` and normalized for iOS client generation.
    The backend exposes JWT bearer authentication via `/api/auth/login` and `/api/auth/refresh`.
    Password reset and email verification flows are also described so the spec can be used for iOS client generation.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/customers/profile:
    get:
      tags: [customers]
      operationId: getCustomerProfile
      summary: Get current customer profile
      responses:
        '200':
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [customers]
      operationId: upsertCustomerProfile
      summary: Create or update current customer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerProfileUpdate'
            example:
              phone: "+79998887766"
              default_address: "Москва, ул. Арбат, 10"
              default_coordinates:
                lat: 55.7525
                lon: 37.5929
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/customers/stores:
    get:
      tags: [customers]
      operationId: listCustomerStores
      security: []
      summary: List active stores, optionally filtered by delivery zone point
      parameters:
        - $ref: '#/components/parameters/LatQuery'
        - $ref: '#/components/parameters/LonQuery'
      responses:
        '200':
          description: Active stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StorePublicResponse'
              example:
                - id: 1
                  name: Bella Italia
                  address: "Москва, Арбат, 18"
                  delivery_zone:
                    type: Polygon
                    coordinates:
                      - - [37.58, 55.76]
                        - [37.61, 55.77]
                        - [37.63, 55.74]
                        - [37.58, 55.76]
                  is_active: true
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/customers/stores/{store_id}/menu:
    get:
      tags: [customers]
      operationId: getCustomerStoreMenu
      security: []
      summary: Get public customer menu for store
      parameters:
        - $ref: '#/components/parameters/StoreID'
      responses:
        '200':
          description: Store menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerMenuResponse'
              example:
                store_id: 1
                categories:
                  - id: 11
                    name: Пицца
                    sort_order: 0
                    items:
                      - id: 101
                        name: Пепперони
                        description: Томатный соус, сыр, пепперони
                        price: "790.00"
                        image_url: https://cdn.example.com/pepperoni.png
        '404':
          $ref: '#/components/responses/NotFound'
  /api/customers/orders:
    post:
      tags: [customers]
      operationId: createCustomerOrder
      summary: Create customer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
            example:
              store_id: 1
              delivery_address: "Москва, ул. Арбат, 10"
              delivery_coordinates:
                lat: 55.7525
                lon: 37.5929
              items:
                - item_id: 101
                  quantity: 2
              comment: "Не звонить в дверь"
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [customers]
      operationId: listCustomerOrders
      summary: List current customer orders
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/customers/orders/{order_id}/status:
    get:
      tags: [customers]
      operationId: getCustomerOrderStatus
      summary: Get customer order status with courier location and ETA
      parameters:
        - $ref: '#/components/parameters/OrderID'
      responses:
        '200':
          description: Order status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusResponse'
              example:
                status: assigned
                courier_location:
                  lat: 55.7601
                  lon: 37.6112
                estimated_time: 25
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/customers/track/{order_id}:
    get:
      tags: [customers]
      operationId: trackCustomerCourier
      summary: Track courier location for order
      parameters:
        - $ref: '#/components/parameters/OrderID'
      responses:
        '200':
          description: Courier location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourierLocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/customers/orders/{order_id}/cancel:
    post:
      tags: [customers]
      operationId: cancelCustomerOrder
      summary: Cancel customer order while still cancellable
      parameters:
        - $ref: '#/components/parameters/OrderID'
      responses:
        '200':
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LatQuery:
      name: lat
      in: query
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    LonQuery:
      name: lon
      in: query
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    OffsetQuery:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    StoreID:
      name: store_id
      in: path
      required: true
      schema:
        type: integer
    OrderID:
      name: order_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Адрес доставки вне зоны доставки магазина
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Недействительный или просроченный токен
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Требуется роль бизнес-пользователя
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Заказ не найден
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationIssue:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
    ValidationErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
...
    OrderStatus:
      type: string
      enum: [draft, pending, confirmed, preparing, ready, assigned, picked_up, on_the_way, delivered, cancelled]
...
    Coordinate:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
    GeoJSONPolygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              minItems: 2
              maxItems: 2
              items:
                type: number
                format: double
...
    CustomerProfileUpdate:
      type: object
      properties:
        phone:
          type: string
          nullable: true
        default_address:
          type: string
          nullable: true
        default_coordinates:
          $ref: '#/components/schemas/Coordinate'
    CustomerProfileResponse:
      allOf:
        - $ref: '#/components/schemas/CustomerProfileUpdate'
        - type: object
          required: [id, user_id]
          properties:
            id:
              type: integer
            user_id:
              type: integer
...
    StorePublicResponse:
      type: object
      required: [id, name, is_active]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        delivery_zone:
          $ref: '#/components/schemas/GeoJSONPolygon'
        is_active:
          type: boolean
    MenuItemPublicResponse:
      type: object
      required: [id, name, price]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
          format: decimal
        image_url:
          type: string
          nullable: true
    CustomerMenuCategoryPublic:
      type: object
      required: [id, name, sort_order, items]
      properties:
        id:
          type: integer
        name:
          type: string
        sort_order:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemPublicResponse'
    CustomerMenuResponse:
      type: object
      required: [store_id, categories]
      properties:
        store_id:
          type: integer
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CustomerMenuCategoryPublic'
    OrderCreateItem:
      type: object
      required: [item_id, quantity]
      properties:
        item_id:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 999
    OrderCreate:
      type: object
      required: [store_id, delivery_address, delivery_coordinates, items]
      properties:
        store_id:
          type: integer
        delivery_address:
          type: string
        delivery_coordinates:
          $ref: '#/components/schemas/Coordinate'
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderCreateItem'
        comment:
          type: string
          nullable: true
    OrderLineSnapshot:
      type: object
      properties:
        item_id:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: string
          format: decimal
        line_total:
          type: string
          format: decimal
    ItemsSnapshot:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineSnapshot'
    OrderResponse:
      type: object
      required: [id, public_id, customer_id, store_id, status, subtotal, delivery_fee, total, created_at, updated_at]
      properties:
        id:
          type: integer
        public_id:
          type: string
        customer_id:
          type: integer
        store_id:
          type: integer
        courier_id:
          type: integer
          nullable: true
        status:
          $ref: '#/components/schemas/OrderStatus'
        delivery_address:
          type: string
          nullable: true
        delivery_coordinates:
          $ref: '#/components/schemas/Coordinate'
        items_snapshot:
          $ref: '#/components/schemas/ItemsSnapshot'
        subtotal:
          type: string
          format: decimal
        delivery_fee:
          type: string
          format: decimal
        total:
          type: string
          format: decimal
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderStatusResponse:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
        courier_location:
          $ref: '#/components/schemas/Coordinate'
        estimated_time:
          type: integer
          nullable: true
    CourierLocationResponse:
      type: object
      required: [id, courier_id, recorded_at]
      properties:
        id:
          type: integer
        courier_id:
          type: integer
        recorded_at:
          type: string
          format: date-time
        coordinates:
          $ref: '#/components/schemas/Coordinate'
        geom_wkt:
          type: string
          nullable: true
...